# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trading bot API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Me",
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/totp/enroll": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Totp Enroll",
        "description": "Generate a fresh TOTP secret and return it + a provisioning URI.\n\nThe secret is NOT persisted yet. The client must immediately call\nPOST /api/me/totp/verify with the secret AND a valid code to\ncommit it.",
        "operationId": "totp_enroll_api_me_totp_enroll_post",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TOTPEnrollResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/totp/verify": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Totp Verify",
        "description": "Commit TOTP enrollment after confirming the user can compute codes.",
        "operationId": "totp_verify_api_me_totp_verify_post",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/totp": {
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Totp Disenroll",
        "description": "Remove TOTP for the current user. Requires a valid current code.",
        "operationId": "totp_disenroll_api_me_totp_delete",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TOTPDisenrollRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/status": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Bot Status",
        "operationId": "bot_status_api_status_get",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Read Config",
        "operationId": "read_config_api_config_get",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigPolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Write Config",
        "operationId": "write_config_api_config_put",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPolicyPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigPolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/active-asset": {
      "get": {
        "tags": [
          "asset"
        ],
        "summary": "Read Active Asset",
        "operationId": "read_active_asset_api_active_asset_get",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveAssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "asset"
        ],
        "summary": "Write Active Asset",
        "operationId": "write_active_asset_api_active_asset_put",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveAssetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveAssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kill": {
      "post": {
        "tags": [
          "kill"
        ],
        "summary": "Request Kill",
        "description": "Request the bot to trip its kill switch.\n\n202 Accepted: the request is queued in Redis; the bot polls and\napplies on its next interval (typically within 2 s). Successive\nrequests are idempotent: a re-poll sees the request even if a\nprevious one was already consumed.",
        "operationId": "request_kill_api_kill_post",
        "parameters": [
          {
            "name": "tradingbot_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tradingbot Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BotStatusResponse": {
        "properties": {
          "connection_state": {
            "type": "string",
            "enum": [
              "connected",
              "disconnected",
              "unknown"
            ],
            "title": "Connection State"
          },
          "kill_switch_tripped": {
            "type": "boolean",
            "title": "Kill Switch Tripped"
          },
          "kill_switch_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kill Switch Reason"
          },
          "open_position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpenPositionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "today_pnl": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PnLResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "connection_state",
          "kill_switch_tripped",
          "kill_switch_reason",
          "open_position",
          "today_pnl"
        ],
        "title": "BotStatusResponse",
        "description": "High-level snapshot for the dashboard poll endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Password"
          },
          "totp_code": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totp Code"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "SetActiveAssetRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9.]{0,7}$",
            "title": "Symbol",
            "description": "US ticker symbol, e.g. AAPL, MSFT, BRK.B."
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "SetActiveAssetRequest",
        "description": "Body for PUT /api/active-asset."
      },
      "TOTPDisenrollRequest": {
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "TOTPDisenrollRequest"
      },
      "TOTPEnrollResponse": {
        "properties": {
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "provisioning_uri": {
            "type": "string",
            "title": "Provisioning Uri"
          }
        },
        "type": "object",
        "required": [
          "secret",
          "provisioning_uri"
        ],
        "title": "TOTPEnrollResponse"
      },
      "TOTPVerifyRequest": {
        "properties": {
          "secret": {
            "type": "string",
            "maxLength": 64,
            "minLength": 16,
            "title": "Secret"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "secret",
          "code"
        ],
        "title": "TOTPVerifyRequest"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "totp_enrolled": {
            "type": "boolean",
            "title": "Totp Enrolled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "is_active"
        ],
        "title": "UserResponse"
      },
...
    }
  }
}